safety-text-classifier/requirements.txt
# Core ML Framework - Fixed versions for compatibility
numpy==1.24.4  # Pin exact version to avoid binary incompatibility
jax[cpu]==0.4.28
flax==0.8.4
optax==0.2.2

# Data and Text Processing  
datasets==2.19.1
transformers==4.40.2
sentence-transformers==2.7.0
tokenizers==0.15.2  # Compatible with transformers 4.40.2
pandas==2.0.3

# Experiment Tracking and Visualization
wandb==0.17.0  # More stable version
matplotlib==3.7.5  # Compatible with numpy 1.24.4
seaborn==0.12.2
plotly==5.17.0

# Web Interface and Serving
gradio==4.26.0
streamlit==1.33.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# Evaluation and Metrics
scikit-learn==1.3.2  # Compatible with numpy 1.24.4
scipy==1.10.1  # Compatible with numpy 1.24.4

# Development and Testing
pytest==8.1.1
black==24.4.0
flake8==7.0.0
jupyter==1.0.0

# Cloud and Monitoring
google-cloud-storage==2.16.0
prometheus-client==0.20.0

# Utilities
tqdm==4.66.4
pyyaml==6.0.1
python-dotenv==1.0.1
pathlib2==2.3.7
